追加学習周りを追加

+++ src/utils.jl
@@ -19,8 +19,8 @@ function save_matrix(x::Matrix; filename="log.txt")
 end
 
 
-function mino_to_array(mino::Union{Nothing, Mino})::Matrix{Int32}
-    res = zeros(Int32, 1, 7)
+function mino_to_array(mino::Union{Nothing, Mino})::Matrix{Float32}
+    res = zeros(Float32, 1, 7)
     index = findfirst(m == mino for m in Tetris.TetrisMino.minos)
     if !isnothing(index)
         res[index] = 1
